+++ src/sema/source_activation.ml
@@ -151,6 +151,8 @@ let event_context = function
             p.delimiter_initializer.initializer_owner.global_header
               .declaration_command
         | Parser.Function_declared p -> p.function_header.declaration_command
+        | Parser.Function_position_written p ->
+            p.position_function.function_header.declaration_command
         | Parser.Function_parameter_declared p ->
             p.parameter_function.function_header.declaration_command
         | Parser.Function_parameter_completed p ->